operator: Move ClusterConfig controller logic in `pkg`
So that we have a dedicated go `package` for this controller (and
variable&co do not overlap between controllers)
